Small fixes for Cg util and GlBuffer compatibility.

+++ include/pangolin/glsl.h
@@ -141,7 +141,7 @@ inline GlSlProgram::GlSlProgram()
 inline GlSlProgram::~GlSlProgram()
 {
     // Remove and delete each shader
-    for(int i=0; i<shaders.size(); ++i ) {
+    for(size_t i=0; i<shaders.size(); ++i ) {
         glDetachObjectARB(prog, shaders[i]);
         glDeleteObjectARB(shaders[i]);
     }
